Add separate unit test class
